+++ train_total_3.sh
@@ -6,7 +6,7 @@ CUDA_VISIBLE_DEVICES=0,1,2,3,4,5,6,7 \
 PORT=25003 \
 USE_MPS=0 \
 ./run.sh \
-./projects/configs/baselines/shape_guide_128_dice.py \
+./projects/configs/baselines/shape_guide_128_dice_weight_full.py \
 8 \
---resume /NHNHOME/WORKSPACE/0526040009_A/hwanhee/Autonomous_Driving_26_ksh_occ_shape_codeonly/work_dirs/shape_guide_128_dice/latest.pth
+--resume /NHNHOME/WORKSPACE/0526040009_A/hwanhee/Autonomous_Driving_26_ksh_occ_shape_codeonly/work_dirs/shape_guide_128_dice_weight_full/latest.pth
 # --cfg-options load_from=/home/hwanhee/Autonomous_Driving_26_0528/epoch_15_lss_only.pth